+++ exporters/otlp/otlp_integration_test.go
@@ -27,8 +27,8 @@ import (
 
 	commonpb "go.opentelemetry.io/otel/exporters/otlp/internal/opentelemetry-proto-gen/common/v1"
 	metricpb "go.opentelemetry.io/otel/exporters/otlp/internal/opentelemetry-proto-gen/metrics/v1"
+	"go.opentelemetry.io/otel/label"
 
-	"go.opentelemetry.io/otel/api/kv"
 	"go.opentelemetry.io/otel/api/metric"
 	metricapi "go.opentelemetry.io/otel/api/metric"
 	"go.opentelemetry.io/otel/exporters/otlp"
@@ -91,15 +91,15 @@ func newExporterEndToEndTest(t *testing.T, additionalOpts []otlp.ExporterOption)
 	}
 	tp1, err := sdktrace.NewProvider(append(pOpts,
 		sdktrace.WithResource(resource.New(
-			kv.String("rk1", "rv11)"),
-			kv.Int64("rk2", 5),
+			label.String("rk1", "rv11)"),
+			label.Int64("rk2", 5),
 		)))...)
 	assert.NoError(t, err)
 
 	tp2, err := sdktrace.NewProvider(append(pOpts,
 		sdktrace.WithResource(resource.New(
-			kv.String("rk1", "rv12)"),
-			kv.Float32("rk3", 6.5),
+			label.String("rk1", "rv12)"),
+			label.Float32("rk3", 6.5),
 		)))...)
 	assert.NoError(t, err)
 
@@ -109,11 +109,11 @@ func newExporterEndToEndTest(t *testing.T, additionalOpts []otlp.ExporterOption)
 	m := 4
 	for i := 0; i < m; i++ {
 		_, span := tr1.Start(context.Background(), "AlwaysSample")
-		span.SetAttributes(kv.Int64("i", int64(i)))
+		span.SetAttributes(label.Int64("i", int64(i)))
 		span.End()
 
 		_, span = tr2.Start(context.Background(), "AlwaysSample")
-		span.SetAttributes(kv.Int64("i", int64(i)))
+		span.SetAttributes(label.Int64("i", int64(i)))
 		span.End()
 	}
 
@@ -124,7 +124,7 @@ func newExporterEndToEndTest(t *testing.T, additionalOpts []otlp.ExporterOption)
 
 	ctx := context.Background()
 	meter := pusher.Provider().Meter("test-meter")
-	labels := []kv.KeyValue{kv.Bool("test", true)}
+	labels := []label.KeyValue{label.Bool("test", true)}
 
 	type data struct {
 		iKind metric.Kind