+++ exporters/stdout/metric_test.go
@@ -26,9 +26,9 @@ import (
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 
-	"go.opentelemetry.io/otel/api/kv"
 	"go.opentelemetry.io/otel/api/metric"
 	"go.opentelemetry.io/otel/exporters/stdout"
+	"go.opentelemetry.io/otel/label"
 	export "go.opentelemetry.io/otel/sdk/export/metric"
 	"go.opentelemetry.io/otel/sdk/export/metric/aggregation"
 	"go.opentelemetry.io/otel/sdk/export/metric/metrictest"
@@ -48,7 +48,7 @@ type testFixture struct {
 	output   *bytes.Buffer
 }
 
-var testResource = resource.New(kv.String("R", "V"))
+var testResource = resource.New(label.String("R", "V"))
 
 func newFixture(t *testing.T, opts ...stdout.Option) testFixture {
 	buf := &bytes.Buffer{}
@@ -145,7 +145,7 @@ func TestStdoutCounterFormat(t *testing.T) {
 	aggregatortest.CheckedUpdate(fix.t, cagg, metric.NewInt64Number(123), &desc)
 	require.NoError(t, cagg.SynchronizedMove(ckpt, &desc))
 
-	checkpointSet.Add(&desc, ckpt, kv.String("A", "B"), kv.String("C", "D"))
+	checkpointSet.Add(&desc, ckpt, label.String("A", "B"), label.String("C", "D"))
 
 	fix.Export(checkpointSet)
 
@@ -163,7 +163,7 @@ func TestStdoutLastValueFormat(t *testing.T) {
 	aggregatortest.CheckedUpdate(fix.t, lvagg, metric.NewFloat64Number(123.456), &desc)
 	require.NoError(t, lvagg.SynchronizedMove(ckpt, &desc))
 
-	checkpointSet.Add(&desc, ckpt, kv.String("A", "B"), kv.String("C", "D"))
+	checkpointSet.Add(&desc, ckpt, label.String("A", "B"), label.String("C", "D"))
 
 	fix.Export(checkpointSet)
 
@@ -183,7 +183,7 @@ func TestStdoutMinMaxSumCount(t *testing.T) {
 	aggregatortest.CheckedUpdate(fix.t, magg, metric.NewFloat64Number(876.543), &desc)
 	require.NoError(t, magg.SynchronizedMove(ckpt, &desc))
 
-	checkpointSet.Add(&desc, ckpt, kv.String("A", "B"), kv.String("C", "D"))
+	checkpointSet.Add(&desc, ckpt, label.String("A", "B"), label.String("C", "D"))
 
 	fix.Export(checkpointSet)
 
@@ -204,7 +204,7 @@ func TestStdoutValueRecorderFormat(t *testing.T) {
 
 	require.NoError(t, aagg.SynchronizedMove(ckpt, &desc))
 
-	checkpointSet.Add(&desc, ckpt, kv.String("A", "B"), kv.String("C", "D"))
+	checkpointSet.Add(&desc, ckpt, label.String("A", "B"), label.String("C", "D"))
 
 	fix.Export(checkpointSet)
 
@@ -268,7 +268,7 @@ func TestStdoutLastValueNotSet(t *testing.T) {
 	lvagg, ckpt := metrictest.Unslice2(lastvalue.New(2))
 	require.NoError(t, lvagg.SynchronizedMove(ckpt, &desc))
 
-	checkpointSet.Add(&desc, lvagg, kv.String("A", "B"), kv.String("C", "D"))
+	checkpointSet.Add(&desc, lvagg, label.String("A", "B"), label.String("C", "D"))
 
 	fix.Export(checkpointSet)
 
@@ -279,9 +279,9 @@ func TestStdoutResource(t *testing.T) {
 	type testCase struct {
 		expect string
 		res    *resource.Resource
-		attrs  []kv.KeyValue
+		attrs  []label.KeyValue
 	}
-	newCase := func(expect string, res *resource.Resource, attrs ...kv.KeyValue) testCase {
+	newCase := func(expect string, res *resource.Resource, attrs ...label.KeyValue) testCase {
 		return testCase{
 			expect: expect,
 			res:    res,
@@ -290,23 +290,23 @@ func TestStdoutResource(t *testing.T) {
 	}
 	testCases := []testCase{
 		newCase("R1=V1,R2=V2,A=B,C=D",
-			resource.New(kv.String("R1", "V1"), kv.String("R2", "V2")),
-			kv.String("A", "B"),
-			kv.String("C", "D")),
+			resource.New(label.String("R1", "V1"), label.String("R2", "V2")),
+			label.String("A", "B"),
+			label.String("C", "D")),
 		newCase("R1=V1,R2=V2",
-			resource.New(kv.String("R1", "V1"), kv.String("R2", "V2")),
+			resource.New(label.String("R1", "V1"), label.String("R2", "V2")),
 		),
 		newCase("A=B,C=D",
 			nil,
-			kv.String("A", "B"),
-			kv.String("C", "D"),
+			label.String("A", "B"),
+			label.String("C", "D"),
 		),
 		// We explicitly do not de-duplicate between resources
 		// and metric labels in this exporter.
 		newCase("R1=V1,R2=V2,R1=V3,R2=V4",
-			resource.New(kv.String("R1", "V1"), kv.String("R2", "V2")),
-			kv.String("R1", "V3"),
-			kv.String("R2", "V4")),
+			resource.New(label.String("R1", "V1"), label.String("R2", "V2")),
+			label.String("R1", "V3"),
+			label.String("R2", "V4")),
 	}
 
 	for _, tc := range testCases {