+++ sdk/trace/trace_test.go
@@ -25,6 +25,7 @@ import (
 	"time"
 
 	"go.opentelemetry.io/otel/api/global"
+	"go.opentelemetry.io/otel/label"
 
 	"github.com/google/go-cmp/cmp"
 	"github.com/stretchr/testify/assert"
@@ -32,7 +33,6 @@ import (
 	grpccodes "google.golang.org/grpc/codes"
 
 	"go.opentelemetry.io/otel/api/apitest"
-	"go.opentelemetry.io/otel/api/kv"
 	"go.opentelemetry.io/otel/api/trace"
 	apitrace "go.opentelemetry.io/otel/api/trace"
 	otelcodes "go.opentelemetry.io/otel/codes"
@@ -92,7 +92,7 @@ func (ts *testSampler) ShouldSample(p SamplingParameters) SamplingResult {
 	if strings.HasPrefix(p.Name, ts.prefix) {
 		decision = RecordAndSampled
 	}
-	return SamplingResult{Decision: decision, Attributes: []kv.KeyValue{kv.Int("callCount", ts.callCount)}}
+	return SamplingResult{Decision: decision, Attributes: []label.KeyValue{label.Int("callCount", ts.callCount)}}
 }
 
 func (ts testSampler) Description() string {
@@ -285,8 +285,8 @@ func TestSetSpanAttributesOnStart(t *testing.T) {
 	tp, _ := NewProvider(WithSyncer(te))
 	span := startSpan(tp,
 		"StartSpanAttribute",
-		apitrace.WithAttributes(kv.String("key1", "value1")),
-		apitrace.WithAttributes(kv.String("key2", "value2")),
+		apitrace.WithAttributes(label.String("key1", "value1")),
+		apitrace.WithAttributes(label.String("key2", "value2")),
 	)
 	got, err := endSpan(te, span)
 	if err != nil {
@@ -300,9 +300,9 @@ func TestSetSpanAttributesOnStart(t *testing.T) {
 		},
 		ParentSpanID: sid,
 		Name:         "span0",
-		Attributes: []kv.KeyValue{
-			kv.String("key1", "value1"),
-			kv.String("key2", "value2"),
+		Attributes: []label.KeyValue{
+			label.String("key1", "value1"),
+			label.String("key2", "value2"),
 		},
 		SpanKind:               apitrace.SpanKindInternal,
 		HasRemoteParent:        true,
@@ -317,7 +317,7 @@ func TestSetSpanAttributes(t *testing.T) {
 	te := &testExporter{}
 	tp, _ := NewProvider(WithSyncer(te))
 	span := startSpan(tp, "SpanAttribute")
-	span.SetAttributes(kv.Key("key1").String("value1"))
+	span.SetAttributes(label.String("key1", "value1"))
 	got, err := endSpan(te, span)
 	if err != nil {
 		t.Fatal(err)
@@ -330,8 +330,8 @@ func TestSetSpanAttributes(t *testing.T) {
 		},
 		ParentSpanID: sid,
 		Name:         "span0",
-		Attributes: []kv.KeyValue{
-			kv.String("key1", "value1"),
+		Attributes: []label.KeyValue{
+			label.String("key1", "value1"),
 		},
 		SpanKind:               apitrace.SpanKindInternal,
 		HasRemoteParent:        true,
@@ -349,10 +349,10 @@ func TestSetSpanAttributesOverLimit(t *testing.T) {
 
 	span := startSpan(tp, "SpanAttributesOverLimit")
 	span.SetAttributes(
-		kv.Bool("key1", true),
-		kv.String("key2", "value2"),
-		kv.Bool("key1", false), // Replace key1.
-		kv.Int64("key4", 4),    // Remove key2 and add key4
+		label.Bool("key1", true),
+		label.String("key2", "value2"),
+		label.Bool("key1", false), // Replace key1.
+		label.Int64("key4", 4),    // Remove key2 and add key4
 	)
 	got, err := endSpan(te, span)
 	if err != nil {
@@ -366,9 +366,9 @@ func TestSetSpanAttributesOverLimit(t *testing.T) {
 		},
 		ParentSpanID: sid,
 		Name:         "span0",
-		Attributes: []kv.KeyValue{
-			kv.Bool("key1", false),
-			kv.Int64("key4", 4),
+		Attributes: []label.KeyValue{
+			label.Bool("key1", false),
+			label.Int64("key4", 4),
 		},
 		SpanKind:               apitrace.SpanKindInternal,
 		HasRemoteParent:        true,
@@ -385,14 +385,14 @@ func TestEvents(t *testing.T) {
 	tp, _ := NewProvider(WithSyncer(te))
 
 	span := startSpan(tp, "Events")
-	k1v1 := kv.Key("key1").String("value1")
-	k2v2 := kv.Bool("key2", true)
-	k3v3 := kv.Int64("key3", 3)
+	k1v1 := label.String("key1", "value1")
+	k2v2 := label.Bool("key2", true)
+	k3v3 := label.Int64("key3", 3)
 
-	span.AddEvent(context.Background(), "foo", kv.Key("key1").String("value1"))
+	span.AddEvent(context.Background(), "foo", label.String("key1", "value1"))
 	span.AddEvent(context.Background(), "bar",
-		kv.Bool("key2", true),
-		kv.Int64("key3", 3),
+		label.Bool("key2", true),
+		label.Int64("key3", 3),
 	)
 	got, err := endSpan(te, span)
 	if err != nil {
@@ -414,8 +414,8 @@ func TestEvents(t *testing.T) {
 		Name:            "span0",
 		HasRemoteParent: true,
 		MessageEvents: []export.Event{
-			{Name: "foo", Attributes: []kv.KeyValue{k1v1}},
-			{Name: "bar", Attributes: []kv.KeyValue{k2v2, k3v3}},
+			{Name: "foo", Attributes: []label.KeyValue{k1v1}},
+			{Name: "bar", Attributes: []label.KeyValue{k2v2, k3v3}},
 		},
 		SpanKind:               apitrace.SpanKindInternal,
 		InstrumentationLibrary: instrumentation.Library{Name: "Events"},
@@ -431,19 +431,19 @@ func TestEventsOverLimit(t *testing.T) {
 	tp, _ := NewProvider(WithConfig(cfg), WithSyncer(te))
 
 	span := startSpan(tp, "EventsOverLimit")
-	k1v1 := kv.Key("key1").String("value1")
-	k2v2 := kv.Bool("key2", false)
-	k3v3 := kv.Key("key3").String("value3")
+	k1v1 := label.String("key1", "value1")
+	k2v2 := label.Bool("key2", false)
+	k3v3 := label.String("key3", "value3")
 
-	span.AddEvent(context.Background(), "fooDrop", kv.Key("key1").String("value1"))
+	span.AddEvent(context.Background(), "fooDrop", label.String("key1", "value1"))
 	span.AddEvent(context.Background(), "barDrop",
-		kv.Bool("key2", true),
-		kv.Key("key3").String("value3"),
+		label.Bool("key2", true),
+		label.String("key3", "value3"),
 	)
-	span.AddEvent(context.Background(), "foo", kv.Key("key1").String("value1"))
+	span.AddEvent(context.Background(), "foo", label.String("key1", "value1"))
 	span.AddEvent(context.Background(), "bar",
-		kv.Bool("key2", false),
-		kv.Key("key3").String("value3"),
+		label.Bool("key2", false),
+		label.String("key3", "value3"),
 	)
 	got, err := endSpan(te, span)
 	if err != nil {
@@ -464,8 +464,8 @@ func TestEventsOverLimit(t *testing.T) {
 		ParentSpanID: sid,
 		Name:         "span0",
 		MessageEvents: []export.Event{
-			{Name: "foo", Attributes: []kv.KeyValue{k1v1}},
-			{Name: "bar", Attributes: []kv.KeyValue{k2v2, k3v3}},
+			{Name: "foo", Attributes: []label.KeyValue{k1v1}},
+			{Name: "bar", Attributes: []label.KeyValue{k2v2, k3v3}},
 		},
 		DroppedMessageEventCount: 2,
 		HasRemoteParent:          true,
@@ -481,18 +481,18 @@ func TestLinks(t *testing.T) {
 	te := &testExporter{}
 	tp, _ := NewProvider(WithSyncer(te))
 
-	k1v1 := kv.Key("key1").String("value1")
-	k2v2 := kv.Key("key2").String("value2")
-	k3v3 := kv.Key("key3").String("value3")
+	k1v1 := label.String("key1", "value1")
+	k2v2 := label.String("key2", "value2")
+	k3v3 := label.String("key3", "value3")
 
 	sc1 := apitrace.SpanContext{TraceID: apitrace.ID([16]byte{1, 1}), SpanID: apitrace.SpanID{3}}
 	sc2 := apitrace.SpanContext{TraceID: apitrace.ID([16]byte{1, 1}), SpanID: apitrace.SpanID{3}}
 
 	span := startSpan(tp, "Links",
-		apitrace.LinkedTo(sc1, kv.Key("key1").String("value1")),
+		apitrace.LinkedTo(sc1, label.String("key1", "value1")),
 		apitrace.LinkedTo(sc2,
-			kv.Key("key2").String("value2"),
-			kv.Key("key3").String("value3"),
+			label.String("key2", "value2"),
+			label.String("key3", "value3"),
 		),
 	)
 
@@ -510,8 +510,8 @@ func TestLinks(t *testing.T) {
 		Name:            "span0",
 		HasRemoteParent: true,
 		Links: []apitrace.Link{
-			{SpanContext: sc1, Attributes: []kv.KeyValue{k1v1}},
-			{SpanContext: sc2, Attributes: []kv.KeyValue{k2v2, k3v3}},
+			{SpanContext: sc1, Attributes: []label.KeyValue{k1v1}},
+			{SpanContext: sc2, Attributes: []label.KeyValue{k2v2, k3v3}},
 		},
 		SpanKind:               apitrace.SpanKindInternal,
 		InstrumentationLibrary: instrumentation.Library{Name: "Links"},
@@ -532,13 +532,13 @@ func TestLinksOverLimit(t *testing.T) {
 	tp, _ := NewProvider(WithConfig(cfg), WithSyncer(te))
 
 	span := startSpan(tp, "LinksOverLimit",
-		apitrace.LinkedTo(sc1, kv.Key("key1").String("value1")),
-		apitrace.LinkedTo(sc2, kv.Key("key2").String("value2")),
-		apitrace.LinkedTo(sc3, kv.Key("key3").String("value3")),
+		apitrace.LinkedTo(sc1, label.String("key1", "value1")),
+		apitrace.LinkedTo(sc2, label.String("key2", "value2")),
+		apitrace.LinkedTo(sc3, label.String("key3", "value3")),
 	)
 
-	k2v2 := kv.Key("key2").String("value2")
-	k3v3 := kv.Key("key3").String("value3")
+	k2v2 := label.String("key2", "value2")
+	k3v3 := label.String("key3", "value3")
 
 	got, err := endSpan(te, span)
 	if err != nil {
@@ -553,8 +553,8 @@ func TestLinksOverLimit(t *testing.T) {
 		ParentSpanID: sid,
 		Name:         "span0",
 		Links: []apitrace.Link{
-			{SpanContext: sc2, Attributes: []kv.KeyValue{k2v2}},
-			{SpanContext: sc3, Attributes: []kv.KeyValue{k3v3}},
+			{SpanContext: sc2, Attributes: []label.KeyValue{k2v2}},
+			{SpanContext: sc3, Attributes: []label.KeyValue{k3v3}},
 		},
 		DroppedLinkCount:       1,
 		HasRemoteParent:        true,
@@ -619,7 +619,7 @@ func TestSetSpanStatus(t *testing.T) {
 
 func cmpDiff(x, y interface{}) string {
 	return cmp.Diff(x, y,
-		cmp.AllowUnexported(kv.Value{}),
+		cmp.AllowUnexported(label.Value{}),
 		cmp.AllowUnexported(export.Event{}))
 }
 
@@ -956,7 +956,7 @@ func TestRecordError(t *testing.T) {
 				{
 					Name: errorEventName,
 					Time: errTime,
-					Attributes: []kv.KeyValue{
+					Attributes: []label.KeyValue{
 						errorTypeKey.String(s.typ),
 						errorMessageKey.String(s.msg),
 					},
@@ -1003,7 +1003,7 @@ func TestRecordErrorWithStatus(t *testing.T) {
 			{
 				Name: errorEventName,
 				Time: errTime,
-				Attributes: []kv.KeyValue{
+				Attributes: []label.KeyValue{
 					errorTypeKey.String("go.opentelemetry.io/otel/internal/testing.TestError"),
 					errorMessageKey.String("test error"),
 				},
@@ -1088,9 +1088,9 @@ func TestWithResource(t *testing.T) {
 	var te testExporter
 	tp, _ := NewProvider(WithSyncer(&te),
 		WithConfig(Config{DefaultSampler: AlwaysSample()}),
-		WithResource(resource.New(kv.String("rk1", "rv1"), kv.Int64("rk2", 5))))
+		WithResource(resource.New(label.String("rk1", "rv1"), label.Int64("rk2", 5))))
 	span := startSpan(tp, "WithResource")
-	span.SetAttributes(kv.String("key1", "value1"))
+	span.SetAttributes(label.String("key1", "value1"))
 	got, err := endSpan(&te, span)
 	if err != nil {
 		t.Error(err.Error())
@@ -1103,12 +1103,12 @@ func TestWithResource(t *testing.T) {
 		},
 		ParentSpanID: sid,
 		Name:         "span0",
-		Attributes: []kv.KeyValue{
-			kv.String("key1", "value1"),
+		Attributes: []label.KeyValue{
+			label.String("key1", "value1"),
 		},
 		SpanKind:               apitrace.SpanKindInternal,
 		HasRemoteParent:        true,
-		Resource:               resource.New(kv.String("rk1", "rv1"), kv.Int64("rk2", 5)),
+		Resource:               resource.New(label.String("rk1", "rv1"), label.Int64("rk2", 5)),
 		InstrumentationLibrary: instrumentation.Library{Name: "WithResource"},
 	}
 	if diff := cmpDiff(got, want); diff != "" {
@@ -1167,7 +1167,7 @@ func TestSpanCapturesPanic(t *testing.T) {
 	require.Len(t, te.spans, 1)
 	require.Len(t, te.spans[0].MessageEvents, 1)
 	assert.Equal(t, te.spans[0].MessageEvents[0].Name, errorEventName)
-	assert.Equal(t, te.spans[0].MessageEvents[0].Attributes, []kv.KeyValue{
+	assert.Equal(t, te.spans[0].MessageEvents[0].Attributes, []label.KeyValue{
 		errorTypeKey.String("*errors.errorString"),
 		errorMessageKey.String("error message"),
 	})